requirements.txt
fastapi
uvicorn
pydantic
redis
spacy
python-dotenv
psycopg2-binary
openai
python-multipart
